change into splat routes
