+++ app/root.jsx
@@ -10,8 +10,8 @@ import {
 import { storyblokInit, apiPlugin } from "@storyblok/react";
 import Feature from "./components/Feature";
 import Grid from "./components/Grid";
-import Page from "./components/Page";
 import Teaser from "./components/Teaser";
+import Page from "./components/Page";
 
 const components = {
   feature: Feature,
@@ -48,4 +48,3 @@ export default function App() {
     </html>
   );
 }
-
